create function for sorting by Date, create function for remove then add sorted sneacers cards, used 'switch' for comparisons elements by id
